CP-34801: fix tests to work with new version of alcotest-lwt
The vhd stress test isn't run during builds because it takes a long
time to execute. Before it wasn't particularly obvious that it
wasn't being run, so we make it explicit in the test make target.

In addition the test had stopped working after we upgraded alcotest-lwt,
which has been fixed.

Signed-off-by: lippirk <ben.anson@citrix.com>